✨ feat: consolidate space fetching logic into a single API endpoint
Replaced the separate 'from-contract' endpoint with a unified handler
that supports both contract lookups and listing modifiable spaces.
This change simplifies the API structure and improves maintainability.
